services/api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
redis>=5.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0